fix: generate integer/number/boolean as their respective types
Instead of returning them as a string, we render them as their actual types
in the json.

Also honour _some_ of the validation properties from the spec:

- minimum and maximum for number/integer
- minItems and maxItems for arrays
- maxLength for strings

Additionally simplify the code a little by generating the mock data
as an object and serialising it at the end. This also means the final
json data is a little more neatly structured too.

Fixes #4